+++ src/main/java/br/com/cast/avaliacao/controller/CategoryController.java
@@ -3,6 +3,7 @@ package br.com.cast.avaliacao.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -18,6 +19,7 @@ public class CategoryController {
 	private CategoryService categoryService;
 	
 	@GetMapping("/todas")
+	@PreAuthorize("hasAuthority('ROLE_SEARCH_CATEGORY')")
 	public @ResponseBody List<Category> findAllCategories() {
 		return this.categoryService.findAll();
 	}