+++ src/main/java/br/com/cast/avaliacao/controller/CourseController.java
@@ -13,6 +13,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -27,6 +28,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import br.com.cast.avaliacao.model.Course;
+import br.com.cast.avaliacao.repository.projection.SummarizedCourse;
 import br.com.cast.avaliacao.service.CourseService;
 
 @RestController
@@ -35,10 +37,12 @@ public class CourseController {
 	@Autowired
 	private CourseService courseService;
 
+	@SuppressWarnings("unused")
 	@Autowired
 	private MessageSource messageSource;
-
+	
 	@GetMapping
+	@PreAuthorize("hasAuthority('ROLE_SEARCH_COURSE') and #oauth2.hasScope('read')")
 	public Page<Course> findPage(@RequestParam(required = false, defaultValue = "") String description,
 			Pageable pageable) {
 		Page<Course> pagedCourse = this.courseService.findPage(description, pageable);
@@ -46,11 +50,19 @@ public class CourseController {
 	}
 
 	@GetMapping("/todos")
+	@PreAuthorize("hasAuthority('ROLE_SEARCH_COURSE') and #oauth2.hasScope('read')")
 	public @ResponseBody List<Course> findAllCourses() {
 		return this.courseService.findAll();
 	}
+	
+	@GetMapping(params = "resumo")
+	@PreAuthorize("hasAuthority('ROLE_SEARCH_COURSE') and #oauth2.hasScope('read')")
+	public @ResponseBody List<SummarizedCourse> findAllSummaryCourses() {
+		return this.courseService.findAllSummarizedCourse();
+	}
 
 	@GetMapping("/{id}")
+	@PreAuthorize("hasAuthority('ROLE_SEARCH_COURSE') and #oauth2.hasScope('read')")
 	public @ResponseBody ResponseEntity<Course> findCourseById(@PathVariable Long id) {
 		Optional<Course> opcionalCourse = courseService.findById(id);
 
@@ -58,6 +70,7 @@ public class CourseController {
 	}
 
 	@DeleteMapping("/{id}")
+	@PreAuthorize("hasAuthority('ROLE_DELETE_COURSE') and #oauth2.hasScope('write')")
 	@ResponseStatus(HttpStatus.NO_CONTENT)
 	public ResponseEntity<Void> deleteCourse(@PathVariable Long id) {
 		this.courseService.delete(id);
@@ -65,6 +78,7 @@ public class CourseController {
 	}
 
 	@PostMapping("/novo")
+	@PreAuthorize("hasAuthority('ROLE_REGISTER_COURSE') and #oauth2.hasScope('write')")
 	public ResponseEntity<Course> saveCourse(@Valid @RequestBody Course course, HttpServletResponse response) {
 		Course persistedCourse = this.courseService.save(course);
 		URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}").buildAndExpand(course.getId())
@@ -74,10 +88,10 @@ public class CourseController {
 	}
 
 	@PutMapping("/{id}")
+	@PreAuthorize("hasAuthority('ROLE_REGISTER_COURSE') and #oauth2.hasScope('read')")
 	public ResponseEntity<Course> updateCourse(@PathVariable Long id, @Valid @RequestBody Course course,
 			HttpServletResponse response) {
 		Course persistedCourse = this.courseService.update(id, course);
 		return ResponseEntity.status(HttpStatus.CREATED).body(persistedCourse);
-
 	}
 }